+++ lib/models/categories/categories.collection.js
@@ -1,48 +1,12 @@
 'use strict';
 
-class CategoriesCollection  {
+const AnyCollection = require("../mongo.js");
+const CategoriesModel = require("./categories.schema.js");
 
-  constructor(schema) {
-    this.schema = schema;
-  }
+class CategoriesCollection extends AnyCollection {
 
-  /**
-   * gets a list 
-   * @param {object} id / category
-   */
-  read(id) {
-    if (id) {
-      return this.schema.find({ _id: id });
-    } else {
-      return this.schema.find({});
-    }
-  }
-
-  /**
-   * adds a note to the list
-   * @param {object} obj / text, category
-   */
-  create(obj) {
-    const newObj = new this.schema(obj);
-    return newObj.save();
-  }
-
-  update(id, obj) {
-    return this.schema.findByIdAndUpdate(id, obj, {
-      new: true,
-    });
-  }
-
-  /**
-   * deletes item from list based on id match
-   * @param {String} id 
-   */
-  delete(id) {
-    let removeItem = {
-      _id: id,
-    };
-
-    return this.schema.deleteOne(removeItem);
+  constructor() {
+    super(CategoriesModel);
   }
 
 }
