+++ lib/models/products/products.collection.js
@@ -1,49 +1,12 @@
 'use strict';
 
-class ProductsCollection  {
+const AnyCollection = require("../mongo.js");
+const ProductsModel = require("./products.schema.js");
 
-  constructor(schema) {
-    this.schema = schema;
-  }
+class ProductsCollection extends AnyCollection {
 
-  /**
-   * gets a list of notes
-   * @param {object} id / category
-   */
-  read(id) {
-
-    if (id) {
-      return this.schema.find({ _id: id });
-    } else {
-      return this.schema.find({});
-    }
-  }
-
-  /**
-   * adds a note to the list
-   * @param {object} obj / text, category
-   */
-  create(obj) {
-    const newObj = new this.schema(obj);
-    return newObj.save();
-  }
-
-  update(id, obj) {
-    return this.schema.findByIdAndUpdate(id, obj, {
-      new: true,
-    });
-  }
-
-  /**
-   * deletes item from list based on id match
-   * @param {String} id 
-   */
-  delete(id) {
-    let removeItem = {
-      _id: id,
-    };
-
-    return this.schema.deleteOne(removeItem);
+  constructor() {
+    super(ProductsModel);
   }
 
 }
